requirements.txt
paramiko>=2.0.0
pytest_mock